Repositorio-servicio

+++ web_music_player_project-master/src/main/java/com/example/repositorio/ArtistaRepositorio.java
@@ -4,5 +4,5 @@ package com.example.repositorio;
 import com.example.dominio.Artista;
 import org.springframework.data.repository.Repository;
 
-public interface ArtistaRepositorio extends Repository<Artista,Long> {
+public interface ArtistaRepositorio extends Repository<Artista,Integer> {
 }
